% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_lib.R
\name{connect_to_model}
\alias{connect_to_model}
\title{Checks to see if model is available in PRISM}
\usage{
connect_to_model(model_name, address = "localhost:5656")
}
\arguments{
\item{model_name}{tment benefit at that marker value}

\item{address}{Server address. Default is "localhost:5656". Could be an IP address, for example: 122.103.54.12.}
}
\value{
0 for sucess and 1 for error
}
\description{
Checks to see if model is available in PRISM
}
