% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{draw_plots}
\alias{draw_plots}
\title{draws plots generated by the model in R Session}
\usage{
draw_plots(plot_number = NULL)
}
\arguments{
\item{plot_number}{the number of the plot to be rendered}
}
\value{
graphical object in R
}
\description{
draws plots generated by the model in R Session
}
