% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{get_async_results}
\alias{get_async_results}
\title{Retrieves async results}
\usage{
get_async_results(
  model_name = NULL,
  token = NULL,
  api_key = NULL,
  server = NULL
)
}
\arguments{
\item{model_name}{name of the model}

\item{token}{async job token}

\item{api_key}{API key}

\item{server}{Server address. Defaults to PMN.}

\item{bypass_router}{bypass server API router, for debugging purposes}
}
\value{
processed (from JSON to R object result of the call)
}
\description{
Retrieves async results
}
