% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{get_default_input}
\alias{get_default_input}
\title{Returns default PRISM model input}
\usage{
get_default_input(model_name = NULL, api_key = NULL, server = NULL)
}
\arguments{
\item{model_name}{name of the model. If null, it will be set to the last call's value.}

\item{api_key}{API key. If null, it will be set to the last call's value.}

\item{server}{Server name. If null, it will be set to the last call's value}
}
\description{
Returns default PRISM model input
}
