% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_lib.R
\name{get_default_output}
\alias{get_default_output}
\title{Returns default PRISM model output}
\usage{
get_default_output()
}
\value{
0 for sucess and 1 for error
}
\description{
Returns default PRISM model output
}
