% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_lib.R
\name{get_default_setting}
\alias{get_default_setting}
\title{Returns default PRISM model settings}
\usage{
get_default_setting()
}
\value{
0 for sucess and 1 for error
}
\description{
Returns default PRISM model settings
}
