% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{get_plots}
\alias{get_plots}
\title{Retrieves plots generated by the model in R Session}
\usage{
get_plots()
}
\value{
URL of plots
}
\description{
Retrieves plots generated by the model in R Session
}
