% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{handshake}
\alias{handshake}
\title{checks to see if a model is available on the cloud server}
\usage{
handshake(model_name, server = default_server())
}
\arguments{
\item{model_name}{name of the model}

\item{server}{base url of the server}
}
\value{
TRUE if model is available on the server, FALSE otherwise
}
\description{
checks to see if a model is available on the cloud server
}
\examples{
handshake("accept")
}
