% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_lib.R
\name{model_run}
\alias{model_run}
\title{Executes PRISM model}
\usage{
model_run()
}
\arguments{
\item{parms}{required custom parameters for current model}
}
\value{
0 for sucess and 1 for error
}
\description{
Executes PRISM model
}
