% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{prism_call}
\alias{prism_call}
\title{Generic call to PRISM server}
\usage{
prism_call(func, base_url, api_key = NULL, ...)
}
\arguments{
\item{func}{function to call}

\item{base_url}{the url to call}

\item{api_key}{API key}
}
\value{
processed (from JSON to R object result of the call)
}
\description{
Generic call to PRISM server
}
