% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{prism_default_server}
\alias{prism_default_server}
\title{Returns the default server path for PRISM server}
\usage{
prism_default_server()
}
\value{
the default server path for PRISM server
}
\description{
Returns the default server path for PRISM server
}
