% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{prism_get_async_results}
\alias{prism_get_async_results}
\title{Retrieves async results}
\usage{
prism_get_async_results(
  model_name = NULL,
  api_key = "",
  local_server = FALSE,
  bypass_router = FALSE,
  token
)
}
\arguments{
\item{model_name}{name of the model}

\item{api_key}{API key}

\item{local_server}{whether or not the call should be directed to the server on localhost. Default is FALSE.}

\item{bypass_router}{bypass server API router, for debugging purposes}

\item{token}{async job token}
}
\value{
processed (from JSON to R object result of the call)
}
\description{
Retrieves async results
}
