% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{prism_get_plots}
\alias{prism_get_plots}
\title{Retrieves plots generated by the model in R Session}
\usage{
prism_get_plots()
}
\value{
URL of plots
}
\description{
Retrieves plots generated by the model in R Session
}
