% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{prism_hand_shake}
\alias{prism_hand_shake}
\title{Basic checks to see if model is available in PRISM}
\usage{
prism_hand_shake(model_name, server = default_server())
}
\arguments{
\item{model_name}{name of the model}

\item{server}{base url of the server}
}
\value{
TRUE if the model exists, FALSE if not
}
\description{
Basic checks to see if model is available in PRISM
}
