% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{prism_model_run}
\alias{prism_model_run}
\title{Executes PRISM model}
\usage{
prism_model_run(
  model_name = NULL,
  model_input = NULL,
  api_key = NULL,
  server = NULL,
  async = FALSE
)
}
\arguments{
\item{model_name}{name of the model}

\item{api_key}{API key}

\item{async}{should the model be called in async mode?}

\item{input}{required custom parameters for current model}

\item{local_server}{whether or not the call should be directed to the server on localhost. Default is FALSE.}

\item{bypass_router}{bypass server API router, for debugging purposes}
}
\value{
0 for success and 1 for error
}
\description{
Executes PRISM model
}
