% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-settings.R
\name{set_api_key}
\alias{set_api_key}
\title{Set Peer Models Network (PMN) API key}
\usage{
set_api_key(key, overwrite = FALSE, temporary = FALSE)
}
\arguments{
\item{key}{a Peer Models Network API key. For more information on keys see the \href{http://www.peermodelsnetwork.com}{Peer Models Network}}

\item{overwrite}{Option to overwrite any existing PMN keys already stored locally.}

\item{temporary}{Option to not store thew API key for use across sessions.}
}
\description{
API key to communicate with models hosted on the Peer Models Network.
}
\examples{
\dontrun{
set_api_key("YOUR_PMN_API_KEY")

# This will set the key permanently until overwritten again
set_api_key("YOUR_PMN_API_KEY")
}
}
