% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client_lib.R
\name{get_default_input}
\alias{get_default_input}
\title{Returns default PRISM model input}
\usage{
get_default_input()
}
\value{
0 for sucess and 1 for error
}
\description{
Returns default PRISM model input
}
