% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{get_default_input}
\alias{get_default_input}
\title{Returns default PRISM model input}
\usage{
get_default_input(
  model_name = NULL,
  api_key = "",
  local_server = FALSE,
  bypass_router = FALSE
)
}
\arguments{
\item{model_name}{name of the model}

\item{api_key}{API key}

\item{local_server}{whether or not the call should be directed to the server on localhost. Default is FALSE.}

\item{bypass_router}{bypass server API router, for debugging purposes}
}
\description{
Returns default PRISM model input
}
