% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{model_run}
\alias{model_run}
\title{Executes PRISM model}
\usage{
model_run(input = NULL)
}
\arguments{
\item{input}{required custom parameters for current model}
}
\value{
0 for success and 1 for error
}
\description{
Executes PRISM model
}
